Set RGB LED digitally for boot logo & flashlight
Instead of PWM, basic digital output is used to control the RGB LED in the
bootLogo() and flashlight() functions. This reduces code size for sketches
which don't use the setRGBled() function. In bootLogo(), instead of dimming
the red LED, the red then green then blue LEDs are lit in sequence at full
brightness.

digitalWriteRGB() was added as a public function and is also used for the
changes above.